requirements.txt
h5py
pyqt5
pyqtgraph
pyyaml
scipy
seaborn
tqdm
imageio
spectral
scikit-learn
matplotlib
opticalglass
pyyaml
snoop
torch
torch-geometric
torch-cluster
